refactor(schemas): migrate Phase 7 endpoint schemas to domain schema files (#6042)
Move locally-defined Pydantic models from 4 endpoint files into domain schema files:
- schemas_system.py: Security assessment schemas (security_assessment.py)
- schemas_knowledge.py: AI stack integration schemas (ai_stack_integration.py)
- schemas_agent.py: Auth + user management schemas (auth.py, user_management/users.py)
